Update build settings and resolve all compiler warnings This commit updates the project to target modern compiler and SDK settings, as well as resolves all existing build warnings.
Key changes include:

Project now targets the latest Windows 10 SDK.

Updated the C++ standard to C++20.

Updated the C standard to C17.# with '#' will be ignored, and an empty message aborts the commit.
	modified:   src/EffectLib/EffectManager.cpp
	modified:   src/EffectLib/EffectMesh.h
modified:   src/EterLib/IME.cpp
	modified:   src/EterLib/MSWindow.cpp
	modified:   src/EterLib/NetStream.cpp
	modified:   src/GameLib/MapBase.h
	modified:   src/GameLib/MapManager.cpp
	modified:   src/GameLib/MapOutdoor.h
	modified:   src/GameLib/MapOutdoorRender.cpp
	modified:   src/SpeedTreeLib/SpeedTreeWrapper.cpp
	modified:   src/UserInterface/MarkManager.cpp
	modified:   src/UserInterface/PythonCharacterManager.cpp
	modified:   src/UserInterface/PythonItem.cpp
	modified:   src/UserInterface/PythonTextTail.cpp
	modified:   src/UserInterface/UserInterface.cpp

+++ src/GameLib/MapManager.cpp
@@ -241,7 +241,7 @@ void CMapManager::BeginEnvironment()
 	{
 		ms_lpd3dDevice->LightEnable(0, TRUE);
 
-		rkMap.ApplyLight((DWORD)mc_pcurEnvironmentData, mc_pcurEnvironmentData->DirLights[ENV_DIRLIGHT_BACKGROUND]);		
+		rkMap.ApplyLight((uintptr_t)mc_pcurEnvironmentData, mc_pcurEnvironmentData->DirLights[ENV_DIRLIGHT_BACKGROUND]);
 	}
 	else
 		ms_lpd3dDevice->LightEnable(0, FALSE);
